AllocArrays

Documentation for AllocArrays.

Public API

AllocArrays.AllocArray — Type

struct AllocArray{T, N,A<:AbstractArray{T,N}} <: AbstractArray{T,N}
    arr::A
end

Wrapper type which forwards most array methods to the inner array arr, but dispatches similar to special allocation methods.

AllocArrays.with_bumper — Function

with_bumper(f)

Runs f() in the context of using a BumperAllocator{Nothing} to allocate memory to similar calls on AllocArrays.

All such allocations should occur within an @no_escape block, and of course, no such allocations should escape that block.

Thread-safe: f may spawn multiple tasks or threads, which may each allocate memory using similar calls on AllocArray's.

Example

using AllocArrays, Bumper
input = AllocArray([1,2,3])
c = Channel(Inf)
with_bumper() do
    Threads.@threads for i = 1:10
        @no_escape begin
            # ...code with may be multithreaded but which must not escape or return newly-allocated AllocArrays...
            put!(c, sum(input .+ i))
        end
     end
     close(c)
end
sum(collect(c))

# output
225

AllocArrays.with_locked_bumper — Function

with_locked_bumper(f, buf::AllocBuffer)

Runs f() in the context of using a LockedBumperAllocator to allocate memory to similar calls on AllocArrays.

All such allocations should occur within an @no_escape block, and of course, no such allocations should escape that block.

Thread-safe: f may spawn multiple tasks or threads, which may each allocate memory using similar calls on AllocArray's.

Example

using AllocArrays, Bumper

buf = default_buffer()
input = AllocArray([1,2,3])
c = Channel(Inf)
with_locked_bumper(buf) do
    # ...code with may be multithreaded but which must not escape or return newly-allocated AllocArrays...
    @sync for i = 1:10
        @no_escape buf begin
            Threads.@spawn put!(c, sum(input .+ i))
        end
    end
    close(c)
end
sum(collect(c))

# output
225

AllocArrays.unsafe_with_bumper — Function

unsafe_with_bumper(f, buf::AllocBuffer)

Runs f() in the context of using a BumperAllocator{typeof(buf)} to allocate memory to similar calls on AllocArrays.

All such allocations should occur within an @no_escape block, and of course, no such allocations should escape that block.

Warning

Not thread-safe. f must not allocate memory using similar calls on AllocArray's across multiple threads or tasks.

Example

using AllocArrays, Bumper
using AllocArrays: unsafe_with_bumper

input = AllocArray([1,2,3])
buf = default_buffer()
unsafe_with_bumper(buf) do
     @no_escape buf begin
        # ...code with must not allocate AllocArrays on multiple tasks via `similar` nor escape or return newly-allocated AllocArrays...
        sum(input .* 2)
     end
end

# output
12